Phase out support for Android Gradle Plugin 2, focus on latest major version
