--- .idea/runConfigurations/Plugin__Run_All_Unit_Tests.xml
@@ -1,22 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Plugin: Run All Unit Tests" type="GradleRunConfiguration" factoryName="Gradle">
-    <ExternalSystemSettings>
-      <option name="executionName" />
-      <option name="externalProjectPath" value="$PROJECT_DIR$" />
-      <option name="externalSystemIdString" value="GRADLE" />
-      <option name="scriptParameters" value="" />
-      <option name="taskDescriptions">
-        <list />
-      </option>
-      <option name="taskNames">
-        <list>
-          <option value=":testCommon:test" />
-          <option value=":testAgp2x:test" />
-          <option value=":testAgp3x:test" />
-        </list>
-      </option>
-      <option name="vmOptions" value="" />
-    </ExternalSystemSettings>
-    <method />
-  </configuration>
-</component>
