+++ android-junit5-tests/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/util/ClasspathSplitter.groovy
@@ -1,0 +1,11 @@
+package de.mannodermaus.gradle.plugins.junit5.util
+
+class ClasspathSplitter {
+
+  static String splitClasspath(List<File> dependencies) {
+    return dependencies
+        .collect { it.absolutePath.replace('\\', '\\\\') }
+        .collect { "'$it'" }
+        .join(", ")
+  }
+}
