+++ android-junit5-tests/src/test/resources/de/mannodermaus/gradle/plugins/junit5/testenv.properties
