--- android-junit5-tests/testCommon/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/util/StringUtils.groovy
@@ -1,11 +1,0 @@
-package de.mannodermaus.gradle.plugins.junit5.util
-
-class StringUtils {
-
-  static String splitClasspath(List<File> dependencies) {
-    return dependencies
-        .collect { it.absolutePath.replace('\\', '\\\\') }
-        .collect { "'$it'" }
-        .join(", ")
-  }
-}
