+++ android-junit5/src/main/kotlin/de/mannodermaus/gradle/plugins/junit5/Interop.kt
@@ -1,8 +1,6 @@
 package de.mannodermaus.gradle.plugins.junit5
 
-import com.android.build.gradle.BaseExtension
 import com.android.build.gradle.api.BaseVariant
-import com.android.build.gradle.internal.dsl.CoreProductFlavor
 import com.android.build.gradle.internal.variant.BaseVariantData
 import org.gradle.api.tasks.TaskInputs
 import org.junit.platform.gradle.plugin.FiltersExtension
@@ -26,9 +24,6 @@ fun FiltersExtension.getExcludeClassNamePatterns(): List<String> =
 
 val BaseVariant.variantData: BaseVariantData
   get() = GroovyInterop.baseVariant_variantData(this)
-
-val BaseExtension.safeDefaultConfig: CoreProductFlavor
-  get() = GroovyInterop.baseExtension_defaultConfig(this)
 
 fun TaskInputs.safeProperty(key: String, value: Any?) =
     GroovyInterop.taskInputs_safeProperty(this, key, value)
