+++ build.gradle
@@ -8,7 +8,7 @@ buildscript {
   dependencies {
     classpath "com.jfrog.bintray.gradle:gradle-bintray-plugin:$BINTRAY_PLUGIN_VERSION"
     classpath "com.github.ben-manes:gradle-versions-plugin:$VERSIONS_PLUGIN_VERSION"
-    classpath "com.android.tools.build:gradle:$ANDROID_PLUGIN_3X_VERSION"
+    classpath "com.android.tools.build:gradle:$ANDROID_PLUGIN_VERSION"
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$KOTLIN_VERSION"
     classpath "org.junit.platform:junit-platform-gradle-plugin:$JUNIT_PLATFORM_VERSION"
     classpath "com.github.dcendents:android-maven-gradle-plugin:$DCENDENTS_MAVEN_PLUGIN_VERSION"
