+++ gradle.properties
@@ -1,47 +1,47 @@
 # suppress inspection "UnusedProperty" for whole file
 
 # "AGP 3.x requires Studio 3.0 minimum" - workaround to use it inside IJ
-android.injected.build.model.only.versioned       = 3
+android.injected.build.model.only.versioned = 3
 
 # Artifact configuration (common)
-GITHUB_URL                                        = https://github.com/mannodermaus/android-junit5
-GITHUB_REPO                                       = mannodermaus/android-junit5
-LICENSE_NAME                                      = Apache-2.0
+GITHUB_URL                                  = https://github.com/mannodermaus/android-junit5
+GITHUB_REPO                                 = mannodermaus/android-junit5
+LICENSE_NAME                                = Apache-2.0
 
 # Artifact configuration (plugin)
-PLUGIN_GROUP_ID                                   = de.mannodermaus.gradle.plugins
-PLUGIN_ARTIFACT_ID                                = android-junit5
-PLUGIN_DESCRIPTION                                = Unit Testing with JUnit 5 for Android.
-PLUGIN_RUNTIME_VERSION_NAME                       = 1.0.31-SNAPSHOT
-PLUGIN_RUNTIME_VERSION_NAME_LATEST_STABLE         = 1.0.30
+PLUGIN_GROUP_ID                             = de.mannodermaus.gradle.plugins
+PLUGIN_ARTIFACT_ID                          = android-junit5
+PLUGIN_DESCRIPTION                          = Unit Testing with JUnit 5 for Android.
+PLUGIN_RUNTIME_VERSION_NAME                 = 1.0.31-SNAPSHOT
+PLUGIN_RUNTIME_VERSION_NAME_LATEST_STABLE   = 1.0.30
 
 # Artifact configuration (embedded-runtime)
-RUNTIME_GROUP_ID                                  = de.mannodermaus.gradle.plugins
-RUNTIME_ARTIFACT_ID                               = android-junit5-embedded-runtime
-RUNTIME_DESCRIPTION                               = Mirror of IntelliJ IDEA's embedded JUnit 5 Runtime.
+RUNTIME_GROUP_ID                            = de.mannodermaus.gradle.plugins
+RUNTIME_ARTIFACT_ID                         = android-junit5-embedded-runtime
+RUNTIME_DESCRIPTION                         = Mirror of IntelliJ IDEA's embedded JUnit 5 Runtime.
 
 # Artifact configuration (instrumentation & runner)
-INSTRUMENTATION_GROUP_ID                          = de.mannodermaus.junit5
-INSTRUMENTATION_ARTIFACT_ID                       = android-instrumentation-test
-INSTRUMENTATION_DESCRIPTION                       = Extensions for instrumented Android tests with JUnit 5.
-INSTRUMENTATION_RUNNER_ARTIFACT_ID                = android-instrumentation-test-runner
-INSTRUMENTATION_RUNNER_DESCRIPTION                = Runner for integration of instrumented Android tests with JUnit 5.
-INSTRUMENTATION_VERSION_NAME                      = 0.1.2-SNAPSHOT
-INSTRUMENTATION_VERSION_NAME_LATEST_STABLE        = 0.1.1
-INSTRUMENTATION_MIN_SDK_VERSION                   = 26
+INSTRUMENTATION_GROUP_ID                    = de.mannodermaus.junit5
+INSTRUMENTATION_ARTIFACT_ID                 = android-instrumentation-test
+INSTRUMENTATION_DESCRIPTION                 = Extensions for instrumented Android tests with JUnit 5.
+INSTRUMENTATION_RUNNER_ARTIFACT_ID          = android-instrumentation-test-runner
+INSTRUMENTATION_RUNNER_DESCRIPTION          = Runner for integration of instrumented Android tests with JUnit 5.
+INSTRUMENTATION_VERSION_NAME                = 0.1.2-SNAPSHOT
+INSTRUMENTATION_VERSION_NAME_LATEST_STABLE  = 0.1.1
+INSTRUMENTATION_MIN_SDK_VERSION             = 26
+INSTRUMENTATION_RUNNER_MIN_SDK_VERSION      = 14
 
 # Dependency versions (plugins)
-ANDROID_PLUGIN_2X_VERSION                         = 2.3.3
-ANDROID_PLUGIN_3X_VERSION                         = 3.0.1
-BINTRAY_PLUGIN_VERSION                            = 1.8.0
-SHADOW_PLUGIN_VERSION                             = 2.0.1
-VERSIONS_PLUGIN_VERSION                           = 0.17.0
-ANDROID_MAVEN_PLUGIN_VERSION                      = 3.1.1
-DCENDENTS_MAVEN_PLUGIN_VERSION                    = 2.0
+ANDROID_PLUGIN_VERSION                      = 3.0.1
+BINTRAY_PLUGIN_VERSION                      = 1.8.0
+SHADOW_PLUGIN_VERSION                       = 2.0.1
+VERSIONS_PLUGIN_VERSION                     = 0.17.0
+ANDROID_MAVEN_PLUGIN_VERSION                = 3.1.1
+DCENDENTS_MAVEN_PLUGIN_VERSION              = 2.0
 
 # Dependency versions (languages)
-GROOVY_VERSION                                    = 3.0.0-alpha-1
-KOTLIN_VERSION                                    = 1.2.20
+GROOVY_VERSION                              = 3.0.0-alpha-1
+KOTLIN_VERSION                              = 1.2.20
 
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 # A Note to my forgetful self:
@@ -50,22 +50,22 @@ KOTLIN_VERSION                                    = 1.2.20
 # to always update the Travis CI .yml config, too
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 # Android Environment (common)
-COMPILE_SDK_VERSION                               = android-27
-BUILD_TOOLS_VERSION                               = 27.0.3
-TARGET_SDK_VERSION                                = 27
+COMPILE_SDK_VERSION                         = android-27
+BUILD_TOOLS_VERSION                         = 27.0.3
+TARGET_SDK_VERSION                          = 27
 
 # Android Environment (sample)
-SAMPLE_MIN_SDK_VERSION                            = 14
+SAMPLE_MIN_SDK_VERSION                      = 14
 
 # Dependency versions
-JUNIT_PLATFORM_VERSION                            = 1.0.3
-JUNIT_JUPITER_VERSION                             = 5.0.3
-JUNIT_VINTAGE_VERSION                             = 4.12.2
-JUNIT4_VERSION                                    = 4.12
-SPOCK_VERSION                                     = 1.1-groovy-2.4
-APACHE_COMMONS_VERSION                            = 2.6
-SUPPORT_TEST_LIBRARY_VERSION                      = 1.0.1
-ASSERTJ_VERSION                                   = 3.8.0
-ASSERTJ_ANDROID_VERSION                           = 1.2.0
-MOCKITO_VERSION                                   = 2.12.0
-ESPRESSO_VERSION                                  = 3.0.1
+JUNIT_PLATFORM_VERSION                      = 1.0.3
+JUNIT_JUPITER_VERSION                       = 5.0.3
+JUNIT_VINTAGE_VERSION                       = 4.12.2
+JUNIT4_VERSION                              = 4.12
+SPOCK_VERSION                               = 1.1-groovy-2.4
+APACHE_COMMONS_VERSION                      = 2.6
+SUPPORT_TEST_LIBRARY_VERSION                = 1.0.1
+ASSERTJ_VERSION                             = 3.8.0
+ASSERTJ_ANDROID_VERSION                     = 1.2.0
+MOCKITO_VERSION                             = 2.12.0
+ESPRESSO_VERSION                            = 3.0.1
