+++ settings.gradle
@@ -2,11 +2,7 @@ rootProject.name = "android-junit5-root"
 
 include ":android-junit5"
 include ":android-junit5-embedded-runtime"
+include ":android-junit5-tests"
 include ":sample"
 include ":instrumentation"
 include ":instrumentation-runner"
-
-["testCommon", "testAgp2x", "testAgp3x"].each { name ->
-  include ":$name"
-  project(":$name").projectDir = file("android-junit5-tests/$name")
-}
